Add scraper for Rite Aid (#475)
This is based on zip code queries for 100 mile radiuses, and keeps a list of the zips that will fairly efficiently cover any state in `sources/riteaid/zip-codes.js`. I calculated these by hand for small states where one or two zips covers them, plus Washington, which is a priority for us. I got some pointers from a friend on calculating an efficient set of zips for larger states, which I'll try and incorporate on a second pass.