Handle community clinics in Albertsons data (#480)
Albertsons started including a community clinic (presumably it’s being operated by a store pharmacy’s staff), which has been causing two errors in production:
- https://sentry.io/organizations/usdr/issues/2813365771
- https://sentry.io/organizations/usdr/issues/2809052505

This solves both of those issues by adding a new brand for “Community Clinic.”

It also fixes a follow-on issue that clinic alerted me to (but which was not unique to it): some pediatric clinics don’t indicate that they are pediatric in the name (but you can tell because they only list pediatric products), which some others have it in the name but list no products. Both of these are edge-casey, but do exist in the actual data.