Alaska has shingles (vaccines) (#487)
This got a lot more complex than expected. Alaska is now including a shingles (zoster) vaccine in PrepMod, so we need to treat it like flu vaccines and ignore it.

While writing a test for that, I realized we weren’t correctly handling locations that only had non-COVID vaccines — we’d still include them, and set `availability` based on the non-COVID slots rather than COVID slots! That’s obviously pretty wrong. So this does a little bit of refactoring to determine whether a schedule is not actually a COVID schedule and skip over it (and its slots) if not, and drop the location if all its schedules are non-COVID ones.