Add manual corrections for some albertsons locations
Some Albertsons locations just have totally malformed data, and not in a way that it's clear we could algorithmically "fix" it like we do when location names are doubled up. This adds the ability to apply manual corrections to locations based on their ID, and adds a correction for Jewel-Osco #3441.