Classify a location as pediatric if *some* vaccines are pediatric
Albertsons has a messy system where pediatric and non-pediatric bookings are sometimes separate entries/URLs. We try and merge these, and to sanity check, we stop and warn if we are trying to merge two adult or two pediatric locations. We've been getting a lot of these warnings intermittently, and the core of the problem really seems to be that, in contrast to when pediatric vaccines first showed up, many pediatric entries now list both adult and pediatric vaccines. This fixes the warnings by classifying a location as pediatric if *some* of the vaccines are pediatric, instead of only if *all* are.

Fixes https://sentry.io/organizations/usdr/issues/2829716008