Keep loading other states even if Rite Aid API fails on one (#531)
We have an issue where the Rite Aid API loader will often have an error on a request for a state’s data and stop, instead of continuing on to try the rest of the states. This handles that, and sends better error info to Sentry to boot.

It also adds rate limiting support.