Load slot-level detail in second pass for Rite Aid scraper (#532)
The Rite Aid web UI now makes *two* requests for each zip code — the first one returns a list of stores, in *mostly* the same format as before. The difference is that the array of slots in it is now empty and it has a `firstAvailableSlot` property. You then have to make a second call to the same endpoint, but with a list of store numbers in the querystring, to get slots filled in.

This code solves for the above and also adds strict validation of the data coming from Rite Aid, so we get big errors and/or notifications if it changes in any way.

Fixes #528.