Add extra checking for Rite Aid update times (#536)
In #529, we discovered Rite Aid had subtly changed the time zone of their datetimes, causing us to store incorrect `valid_at` times. This adds a little extra checking (we should never see an update time in the future or too far in the past) to help defend against similar changes in the future. It also separates the parsing code into a different function, since it's gotten fairly complex.