Ignore Dallas ISD locations with nonsense addresses
We've been logging errors for two Albertsons/MHealth locations with the nonsense addresses:

    ., ., TX, 75211
    ., ., TX, 75244

I left them as an error under the assumption they were a mistyped entry that would get corrected (you can even see it when you search for appointments by zip code in their UI!), but it never did. At this point, I assume it's intentional (Dallas ISD is probably the school district -- Dallas Independent School District -- and I'm guessing these are for tracking vaccinations across the district, so someone probably did this because a particular address was not releveant).

It's filling up our error logs and charts and making it hard to see less common errors, so this change just makes us skip the location before we even try to parse these nonsense addresses.