Ensure locations’ `state` is a two letter postal abbreviation (#598)
This adds some validation logic to the server that ensures we always store states in the form of a two-letter postal abbreviation (e.g. “AK” for Alaska). This is already what we mostly have, but there are a handful of locations that have come in from PrepMod with different names (e.g. “alaska” or “Alaska”; PrepMod appears to allow free-form data in the state field). The basic premise here is to accept any known identifier for the state (it‘s full name, a postal abbreviation, an ISO name, etc.) and replace that with the postal abbreviation, or return a 422 error if we don’t recognize the state that was passed in.

While I was at it, I went ahead and wrote a more generalized `validateLocationInput()` function that does a few other minor validations (it handles one we already had, and adds a few others for safety). It stops short of validating the complete schema of the location input object, but I’ll add that in if anyone thinks we should just go ahead and do it.

This also includes a migration to clean out the existing bad data.

Fixes #444.