Use more robust deduping algorithm from server
This differs *slightly* from the server version, in that it will return groups with only one entry for items that have no dupes (the server version only returns groups of duplicates, i.e. groups with length > 1). If we ever get around to making loader and server share code, we should smooth that out.