Support Chicago Fire Department community clinics (#616)
Albertsons now has a clinic at a Chicago Fire Department location, which uses a number (the firehouse number?), so it doesn't auto-identify as a community clinic. This adds a special case to support it, much like we did with the Teamsters Union hall.

Fixes https://sentry.io/organizations/usdr/issues/2809052505.