HOTFIX: change external_ids primary key to bigint
We've exhausted every possible integer for the primary key to the `external_ids` table, and as a result, and no longer insert or update any rows in it. Changing the type from integer to bigint gives us some breathing room for the system to keep operating, although we should ideally follow this up with a solution that doesn't run through sequence values so fast.

Fixes https://sentry.io/organizations/usdr/issues/3195943236 and https://sentry.io/organizations/usdr/issues/3155794613