Scrape from mains store pages, not pharmacy pages
It turns out there are a handful of locations we've seen in Alberstons's appointment data that don't correspond to stores we picked up from crawling `local.pharmacy.<brand>.com`, so this code now scrapes from `local.<brand>.com` instead and clicks on the pharmacy link. The output format is now:

    {
      ...(scraped data from main store page),
      pharmacy: {
        ...(scraped data from pharmacy page)
      }
    }

This helps us pick up possible Albertsons locations a little more universally. In practice, we're still just using the `pharmacy` key, but at least the scraper can now pull a broader range of locations that we can reformat and choose to use in the appointment-gathering code if we want.

This is the code @astonm identified as missing from 6c4420c2cfdfc2d0c39f67f5a8020276d59b445a.