Add data file for unlisted Albertsons stores
This adds a data file to cover the two unlisted stores discussed in https://github.com/usdigitalresponse/univaf/pull/677#issuecomment-1146755297

It's a `.js` file instead of `.json` so we can have comments describing each case, since these are special. I also did the same for the Haggen file, so it can have a comment at the top describing why it's a separate file and is hand-coded.