Add scraper to find every Albertsons pharmacy (#677)
This is a follow-on to #676, and attempts to resolve ongoing issues with Albertsons location address formats by scraping and storing a list of every Albertsons-owned store, then matching every appointment data location to one of these known stores. This seems to largely and successfully fix the issues.

One side effect here worth noting here is that this revealed a lot of store numbers that are actually incorrect in the appointment data! In some rare occasions, it’s a typo, bust most often they are using an old store numbering system. This fixes that, but this change will be followed with some cleanup so we don’t have any collisions of incorrect and correct external IDs later on. See https://github.com/usdigitalresponse/univaf/pull/677#issuecomment-1149092024 and other comments in PR #677 for more.