Handle new CDC name format for some Walmarts
When Walmarts had a store number in the name (instead of in the store number field), they used to be formatted as "Walmart NNN" (where N is a digit). Some are now appearing with a 10- prefix, e.g. "Walmart 10-0178" (the same format as in the store number field). This alters our parser to handle both.

Fixes https://sentry.io/organizations/usdr/issues/3338103303.