Support new 00069-2025-* NDC codes
New `00069-2025-*` NDC codes started showing up in CDC's data today. They aren't (yet?) listed on CDC's COVID-19 codes page, but are listed on the complete crosswalk of all vaccines. These are for Comirnaty (the sales name for Pfizer's vaccine), so appear to replace the `00069-1000-*`. Those were always described with a big caveat about having been licensed but never manufactured, and that note appears to have been updated with a bit about NDCs for Comirnaty only being planned for actual use with the upcoming tris-sucrose formulation, which matches the description in this new CDC data for `00069-2025-*`. So basically: the old codes for Comirnaty were for a version that never actually got manufactured under that brand name. These codes are for the new, actually manufactured version of Pfizer's Comirnaty branded vaccines, and are finally being used for the first time.

Fixes https://sentry.io/organizations/usdr/issues/3342633674