Support new 00069-2025-* NDC codes (#699)
New `00069-2025-*` NDC codes started showing up in CDC's data today. They aren't (yet?) listed on CDC's COVID-19 codes page, but are listed on the complete crosswalk of all vaccines. These are for Comirnaty (the sales name for Pfizer's vaccine), so appear to effectively replace the `00069-1000-*` codes.

(More technically: `0069-1000-*` was for a version of the vaccine that was licensed but never manufactured, and we don’t appear to have ever seen it used in practice. `00069-2025-*` is a the newer tris-sucrose formulation of the same vaccine, which appears to finally be getting manufactured and sold now.)

Fixes https://sentry.io/organizations/usdr/issues/3342633674