Test whether Terraform state is good
This is a test of what Terraform Cloud is currently trying to do. #703 failed to deploy because the DNS record it added already exists, which seems reasonable. This is a test to see whether things are all good now, or if it will be stuck until I delete that record and let Terraform recreate it.