HOTFIX: Don't redirect healthcheck to primary host (#705)
It turns out a redirect on the healthcheck endpoint is a failure, since it's not a 200 status code. Without this fix, ECS won't deploy our code, since it thinks the server is failing to run. :(

While I was at it, I also moved DataDog earlier in the stack, so we track errors that happen in middleware.