HOTFIX: Disable primary host redirects (#706)
The primary host redirects we just added worked great in my testing, but do not seem to be playing nice with CloudFront and/or ELB (or maybe it’s because we have two levels of proxying there…?). This disables the redirects and adds logging in their place to see what’s going on.