Ensure www redirect goes to https protocol
The `www` → (no subdomain) redirect was previously switching the protocol to HTTP, resulting in multiple redirects:

    https://www.getmyvax.org/x
      → http://getmyvax.org/x
        → https://getmyvax.org/x

I'm pretty sure this was happening because Cloudfront is terminating the SSL connection, so by the time the request gets to the ELB (where the redirect logic lives) it's seeing an HTTP request instead of HTTPS, and defaults to redirecting whatever protocol was received. This change just explicitly tells it to redirect to HTTPS regardless of what came in.