Cleanup after `www` redirect mess (#711)
Adding redirects from `www.getmyvax.org` to `getmyvax.org` in #665 got a little messy and involved a whole mess of follow-on fixes. This cleans up all the mess that was left behind by them.

This also fixes a minor issue where the `www` → (no subdomain) redirect was switching the protocol to HTTP, resulting in multiple redirects:

    https://www.getmyvax.org/x
      → http://getmyvax.org/x
        → https://getmyvax.org/x

(This was because the redirect happens in the ELB, but the SSL connection is terminated at Cloudfront. So the redirect logic was seeing an HTTP request and reflecting that in the redirect.)