Add constants and codes for early childhood (#714)
Vaccines for ages 6 months - 4 years (Pfizer) and 6 months - 5 years (Moderna) were approved today, and this adds constants for them as well as support for their known NDC and CVX codes so we are likely to be able to recognize them in the data.

It also renames `pfizer_age_2_4` to `pfizer_age_0_4` for clarity (`pfizer_age_2_4` was never actually used, and we checked with partners to make sure everyone was able to handle the change) and tightens our fuzzy matching code to be a little narrower so we get alerts for really unexpected stuff.

This also switches HyVee from using fuzzy matching to exact matching for vaccine types. They clearly have well-defined objects on their side, so it seems like this is what we should have always been doing.

Covers everything I think we can reasonably do in advance on #710.