Split metadata from address on last ` - ` (#700)
This is an attempt to rethink and improve address parsing for Albertsons. When we started, Albertsons data looked pretty reliably like `Safeway 0187 - 1539 Whatever Rd, Somewhere, NJ 08022`, so we split the location name and address on the first dash with a regular expression. These days, more and more data, often dash-separated, has been stuffed into the address field, and we often wind up with other metadata in the address when we parse this way (a complicated new address might be `Albertsons 4406 - Pediatric Vaccine Clinic  - Albertsons - 11330 N W 51st Ave, Gig Harbor, WA, 98332`, which obviously does not parse well with the previous approach.