Use the standard loader logger everywhere
We wrote a decent logging abstraction that handles a lot of cases for the loaders, but never updated all the loaders to use it. This does so. There is probably some good refactoring to be done on it, but that'll be easier after making everything use the existing logging tools.