Handle early childhood Moderna from Albertsons
This is pretty messy, because this change revealed that Albertsons is using `ModernaChild` as a code for early childhood Moderna, but is using `PfizerChild` as a code for *both* early childhood (6 months - 4 years) and normal pediatric (5-11 years) Pfizer. Locations that are only for early childhood Pfizer have "infant" in the address.

Worth noting: it's not entirely clear whether this is totally correct. Depending how you arrive at a booking URL for `PfizerChild`, it may allow you too choose either "3-4 years old" or "5-11 years old". It may also start with a full page notice that talks only about "5-11 years old" but then still allow you to pick both on the next screen. So it's not 100% how exactly we should treat `PfizerChild`.