Match NJVSS locations with different providers
In https://sentry.io/organizations/usdr/issues/3373843306, we're getting a warning because the NJVSS loader can't match an NJVSS location to one we've previously saved. It turns out this is because it has a different provider (`walmart`) which is pretty reasonable. The real problem is that we were only considering locations with NJVSS as a provider for matching, when we should instead be considering all locations in New Jersey.