Handle early childhood Moderna from Albertsons (#734)
Albertsons added a new code (`ModernaChild`) for Moderna’s early childhood vaccine (ages 6 mo – 5 yr) and this adds support for it.

However, they also started using `PfizerChild` for both Pfizer's early childhood vaccine (ages 6 mo – 4 yr) and the mid-childhood vaccine (ages 5–11). There *is* some differentiating info in the location address (if it says “infant,” it’s for the early childhood vaccine), so we try and use that to determine which vaccine(s) are actually on offer. (The reality is that the way things work on their booking site is pretty confusing, and it’s not clear what the right answer is for this. It’s possible they may not even have a good handle on it themselves.)

This also changes how we provide detailed booking URLs. There used to be a set of `meta.booking_url_<adult|pediatric|infant>` properties, but they no longer effectively capture all the booking links. This change adds `meta.booking_urls`, which is meant to replace the older links. It is an array of objects like `{ url: "", products: [], available: "YES"|"NO" }`.