Extend availability merge window form 2 days to 7
It turns out we have a lot of locations whose CDC data is dropping out of our availability listings because reporting frequency has dropped over time — they only report every few days, and the CDC only updates the public data once a day. Since a lot of locations don’t provide product info in their appointment listings, the CDC is a critical source for the `products` field and it makes sense to extend the timeframe we consider valid to contribute to the merged `availability` object.

Fixes #739.