Handle H-E-B's new "Multiple" vaccine type
H-E-B now lists some appointment slots as having a vaccine type of "Multiple," instead of specifying a particular vaccine. There doesn't appear to be anything in the main listing file that includes details as to which vaccines are available, but we can usually get the list of products from the CDC in these cases instead, so for now we just skip this type.

Fixes https://sentry.io/organizations/usdr/issues/3387020042