Support new Moderna `80777-275-*` NDC codes
These codes appear to be the first case where a product can be used for different age groups (albeit in different use cases for those groups), which poses a bit of a problem. I've classified this as `moderna_age_6_11` since it's the product for the primary series as opposed to the booster in that age group.

Fixes https://sentry.io/organizations/usdr/issues/3342633674