Use same DB connection pool for data and logs
We used to create two connection pools to the database (one for data and one for writing availability update logs). This leads to a *lot* of connections, and on the new Render deployment we are sometimes running out of connections and dropping logs. This re-uses the same pool for both, which should be safe (we still keep separate references, so the availability log can be changed to point somewhere else) and solves the issue by using half as many connections.