Log changes to database connection pool size (#781)
While addressing https://sentry.io/organizations/usdr/issues/3155794613, we realized we don't really have a handle on exactly how the database connection pools are behaving. This change logs every time the pool grows or shrinks, which should give us a better idea how the two pools are behaving. It'll log lines like:

```
2022-07-28T19:37:00.711Z [univaf] info: squarpix.lan Data DB pool size: 1
2022-07-28T19:37:00.714Z [univaf] info: squarpix.lan Data DB pool size: 2
2022-07-28T19:37:00.715Z [univaf] info: squarpix.lan Data DB pool size: 3
2022-07-28T19:37:31.763Z [univaf] info: squarpix.lan Data DB pool size: 2
```

(Note: the hostname is included here because we are usually running multiple instances of the server, and we need to know which instance we are logging pool info for to actually understand things.)