Handle Moderna ages 12-17 listings from HyVee
HyVee started listing "Moderna (12-17)" as a vaccine type, but it's not actually a different product from what's used for ages 12+ or 18+. The location in question also lists some younger pediatric vaccines, so presumably this doesn't have anything to do with the minimum age the pharmacist is qualified for, and presumably there are just some appointment slots that are age limited this way. I've classified these as normal `moderna` here, and that is probably a close enough approximation to make sense, rather than creating a new vaccine type.