Graceful shutdown on more signals
We used to shut down the server gracefully on SIGINT, but not SIGTERM. However, SIGTERM is what shuts us down in production, which is arguably the time at which we need to be more graceful! This sets up graceful shutdown for several relevant signals.

It also makes sure we destroy availability logging connections during shutdown (not just normal data connections). This isn't a *huge* deal, since these connections would get reaped pretty quickly when the process dies anyway, but this should make things a little nicer on the DB.