Check for and handle errors in HyVee responses
It turns out we weren't checking for and surfacing error messages from HyVee's GraphQL responses, which led to this confusing error early this morning: https://sentry.io/organizations/usdr/issues/3505575050. We didn't capture much detail because we didn't even realize the response was an error. Now we surface a better error at a more appropriate point.

This is a standard format for errors in GraphQL, so I've just genericized the handling we already had in the Washington Dept. of Health source.