Remove ability to run loader as HTTP server (#841)
Back when we were working with Vaccinate the States, we implemented the ability to run the loader as a server. You could send it an HTTP POST request with a JSON body equivalent to the normal CLI arguments, and it would load data from the requested sources just as if you had run the loader normally from the CLI. The goal was to make it easy to run on Google Cloud, which has a mechanism for scheduling HTTP requests but not for scheduling scripts or Docker runs.

That service only ran briefly, and Vaccinate the States has since shut down. There's not much of a useful reason to retain this functionality, and removing it cleans out a surprising alternate mode for maintainers to worry about.