Add support for Jynneos in Alaska
Folks in Alaska have asked for information on Jynneos (the Monkeypox vaccine) to be included from the state's PrepMod instance, so this adds support. It also adds support in the CDC source, although the actual data doesn't (currently?) show Jynneos yet (and there doesn't seem to be another dataset that does; who knows what's going to happen with that).