Add `MOBILE_CLINIC` location type (#883)
When I was reviewing #122 earlier this week, I noticed some locations had a `MOBILE_CLINIC` location type in the database, even though we don't have constants here for them (they were probably manually set, or came from earlier code, or something to do with Vaccinate the States). This adds a constant for it so we can use it, and so it's clear that it's a valid value, even though we don't actually have any code using it right now.