Simplify DB connection config on Render
One discovery from initially deploying to render in #755 was that database connection configuration is a lot of boilerplate, since we don't support a connection URI as a single variable. This fixes that problem. I've updated the Render configuration here, but not bothered with Terraform/AWS since we will be tearing that down shortly.