Reduce resources for daily data dump job (#904)
I did some quick experiments on Render with a temporary cron job and discovered the bottleneck for the Daily Data Snaphsot is really uploading to S3. There might be some optimization we can do there, but in the mean time, there's no reason to use 2GB of RAM and 1 CPU when less will make no difference in performance.