Try downsizing the database by one level in Render (#905)
While reviewing the Render deployment with @alexallain last week, he pointed out that the database, while smaller than what we have on AWS, may still be overprovisioned (we use a good deal of RAM, but Postgres allocates a huge chunk by default even when it may not actually be using it, so good stats are more complicated here). If we're going to see whether we can lower the database tier by one, we ought to do it now, before switching production to Render.