Try downsizing Render database, take 2 (#908)
This is complicated, since #905 did not work. What I did was disconnected the blueprint on the `main` branch, manually created a smaller DB, imported a fresh dump of production data, updated the `render.yaml` to match new DB, and added `render.yaml` back as the blueprint. Everything sync’d up to the new DB just right. (Did this all on a branch, but as of merging this PR, I’ll repeat the disconnect and reconnect steps to get it sync’d to `main` again.)