Ensure Sentry releases are set on Render Loader (#911)
Use `RENDER_GIT_COMMIT` to set the release tag for Sentry, just like we do in the server. (Ideally, there’d be shared code here, but we don’t have a good way to do that yet.)

This also adds CLI arguments as context in Sentry, since they are helpful in identifying what things are running when an error or warning is captured.