Disallow public connections to the DB on Render
In preparation for switching production to Render we are turning off public internet access to the database. This mainly means you can no longer run `psql -d xyz` from your local machine, and will instead need to start a temporary shell or ssh session on one of our Render services to access it.