If there's no route, tag metrics with the URL path
While reviewing the improved metrics from #915 earlier today, I noticed that a lot of external requests have no route, which turns out to be because the static file server doesn't count as a route (or routes). This attempts to remedy the problem by filling the requested URL (minus the querystring) for the route when the URL actually returned something.