Point `render.getmyvax.org` to Render (#933)
Since my previous attempt (#930, #932) at making CloudFront point getmyvax.org at Render failed spectacularly, I'm trying this again on a separate test domain (render.getmyvax.org) so we don't have downtime. I think the main issue is just that Render needs DNS pointed directly at it first so it can validate that it's ok to respond to the hostname we're sending it.

Once Render validates, this will need to be followed up my a change to the Route53 record so it points to CloudFront instead.