Add more detailed tracing for Sentry
This uses Sentry's tracing support to add some more spans around various operations (mostly in the update path) so we get more detailed traces. There's a lot "seems like it should be built in" utility tooling I had to write here, which was a bit disappointing. And I'm not 100% sure this was totally worthwhile.