Fix broken autoscaling configuration in Terraform
This is pretty subtle! After turning off the loaders in AWS today (#941), I expected the API to scale down to its minimum, but it did not. At first I thought something must somehow still be sending it data or reaching it for queries, but no, it's actually because we had a lower bound instead of an upper bound set on the downscaling (clearly a copy-paste error). This fixes it and also updates the deprecated configurations so someone referencing them won't make the same mistake.