Move nightly snapshot closer to midnight UTC
Out of an abundance of caution, we've been waiting until 5am UTC to archive the previous day's data. That's probably a lot longer than we really need to wait, though, so this moves it up to 1am UTC, making our archived data in S3 show up in a more timely fashion and (for everything except logs) more accurately reflect what the DB looked like at the end of the day.