Fix a DataDog agent issue with ECS_FARGATE hack (#947)
Running the DataDog agent as a separate service (currently recommended by Render) comes with a bunch of caveats, including that it sends a whole raft of system metrics that are pointless, since nothing except the agent is actually run on the system it's gathering metrics for. Making the agent think it's on ECS Fargate using the `ECS_FARGATE` environment variable causes it not to send those.