Stop building and deploying Docker images to ECR (#942)
Part of #683/#779.

Since our Render deployment is not Docker-based and we are shutting off things in AWS ECS, it no longer makes sense to build and publish Docker images. This does not remove our Docker files, both so it’s not too hard to spin things back up again if we have to switch back to ECS, and since they are referenced in our README as part of one kind of local development workflow (even though it’s one I don’t use anymore).