Support bivalent vaccines at H-E-B (#969)
H-E-B added bivalent vaccine listings in an unenjoyably complex new way, and this adds support for it. Fixes https://sentry.io/organizations/usdr/issues/3613330687 and https://sentry.io/organizations/usdr/issues/3613330594.

Previously, each locations would have a `slotDetails` field with an array of objects representing the number of open slots for each vaccine type (identified by the `manufacturer` field [side note: this originally indicated just the manufacturer, until we had pediatric vaccines, at which point it became more of a product type]). With bivalent boosters, they didn’t create a new `manufacturer` value, but instead added a new location-level `availableImmunizations` field that the `manufacter` field maps to (the actual concrete values don’t match, so I had to guess the mappings, but they are pretty obvious). A `manufacturer` like "pfizer" maps to multiple immunizations, like "COVID-19 Pfizer" and "COVID-19 Pfizer_Updated_Booster". To add to the complexity, `availableImmunizations` is often null.

This attempts to map `manufacturer` to `availableImmunizations` when present, or to the list of known values for `availableImmunizations` when not. One upshot here is that we can often list the actual products in the `Multiple` case, which we could not do before. It’s definitely a lot more complicated, though.