Upgrade Render job resources
This week, a lot of our cron jobs on Render started failing due to some sort of scheduling issues in Render's orchestration (that is: the jobs are never actually starting and our scripts don't even get a chance to fail -- the orchestrator is scheduling a job, and then failing it several minutes later instead of running it).

This is a check to see if maybe committing more resources to these jobs can solve the problem.