Washington DoH source is in bad shape (#1008)
The WA Dept. of Health source appears to have stopped finding new data sometime after October 1, which I only noticed as part of cleaning up all the cron problems we were having on Render.

This does a variety of cleanup tasks to fix the issue, account for other changes that appear to have happened while we were no longer finding locations for a few days, and alert us if something similar happens again.