AWS again: reinstate API service on ECS (#1016)
Rebuilds the API service on ECS. This involves a lot of fiddly bits with security groups and policies that I modified *slightly* to see how necessary they are (things may break). There's also a *lot* of room for streamlining and refactoring to clean these up. I'll leave *that* for a follow-on change, though.