AWS again: reinstate the loaders
This sets up the loaders. I'm trying something new here by using `for_each` to reduce boilerplace for each loader definition. If this works well, I think it probably suggest we should rewrite the `loader` module as a `cron` module that can also be used for the Daily Data Snapshot task (i.e. make it about running a task on a schedule, not about laoders specifically).