Change names for metrics that were distributions (#1029)
In #1028, I changed the loader to use the Datadog API instead of the agent. However, the API client we're using doesn't support distributions (yet) and I foolishly used distributions for some of the metrics. I changed their types when switching to the API, but it turns out that while you can change the type of counts/gauges/rates, you can't change between those and distributions. So I'm giving the things that were distributions new names.