Bump the number of DB connections allowed
Allow the API server to use 15 connections at a time on AWS. Looking at our charts, we are generally not bottlenecked on CPU usage, so we should be able to increase performance by letting the server process more requests at a time (by giving it more connections to use). We'll see how this affects our IOPS, though!