Use more smaller API server instances (#1037)
One lesson from rebuilding things on AWS is that we just don't use much memory in the API server. We can't reduce RAM without also cutting the CPU in half, but my working theory is that we can just double the instances and it should be fine. We'll follow the charts and see how this goes (honestly I think a baseline of more instances might make the whole thing more stable anyway).