Switch to ignores in ESLint config file
`.eslintignore` files are no longer supported in ESLint's new-style configuration. We aren't using that yet, but it points the way that using those files will be deprecated (and eventually removed) soon. The new style puts the contents of the ignore files directly in the config file, which is also supported with old-style configs. I've updated to do that here.