Go back to higher CPU instances for API server (#1056)
We don’t need more than a quarter of this memory, but when we use the CPU, we are really using it. Switching to smaller instances appears to have doubled our response latency, so it was probably not a good move. This reverses #1037.

Also add telemetry for how much the loader is retrying failed requests.