Try not setting `throwHttpErrors: false` for Got
I added metrics about retries to the API client a couple days ago, and while we have seen errors that *should* have been retried (e.g. ECONNRESET, ETIMEDOUT), we've never gotten > 0 retries in our metrics. It's possible that when we've seen these errors, they have always occurred for all 3 tries, but that seems unlikely. My quick thought here is that setting `throwHttpErrors: false` is short-circuiting the retry logic (internally, it might be part of the throw-an-exception logic). This might be wrong! It's an experiment.