Actually fix retry issues in ApiClient
In #1061 and #1062, I ran into some dumb issues with retries in the `ApiClient`. After some testing, the problem became obvious: `POST` requests don't retry by default. That's reasonable, and I probably should have assumed that was the problem in the first place, although the got docs have an example of doing replies on a POST without first configuring that to be ok! So I blame this on an incorrect manual, too.