Actually fix retry issues in ApiClient (#1063)
In #1061 and #1062, I ran into some dumb issues with retries in the ApiClient. After some testing, the problem became obvious: POST requests don't retry by default. That's reasonable, and I probably should have assumed that was the problem in the first place, although the got docs have an example of doing replies on a POST without first configuring that to be ok (I even wound up trying this and realized the example is just broken)! So I blame this on an incorrect manual, too. 😉

This also adds tests so I can’t make the same very dumb mistake I made in #1061.