Upgrade Jest to v29.3.1 in /server (#1117)
The `@types/jest` third-party types have been lagging behind, so I went ahead and switched to the official `@jest/globals` method for getting global types. It's a little more verbose (you have to explicitly import the `describe`, `it`, `expect`, etc. functions in each test module), but it the officially supported way, and makes our custom matchers less hacky. It will also always be up-to-date with Jest.