Fix UI deploy workflow
I broke the UI deploy workflow in #1133 - I treated it as if it were installed in the main working directory of the CI server, but it's not (because there are *two* checkouts side-by-side in that workflow). This fixes the path from which we run NPM, which should solve the issue.