Add Docker layer caching in CI
Docker layers are big and include our dependencies, so we should probably be caching them in CI just like we do our dependencies for tests, linting, etc.

This makes use of the `docker/` GitHub actions instead of our own scripts (like we used to do) because they automatically set some of the configuration up for caching with GitHub Actions.